GET request for LinkedBlogs implemented
